Move Subscription and Socket tests to framework package
All of the Subscription and most of the Socket test modules are covering
the generic functionality and do not rely on any specific protocols, so
it makes sense to move them into the framework package where that
functionality lives.

Some complications:

Both socket and subscription tests did slightly rely on node-to-node
things; specifically the protocol versioning. This is realitively easy
to change. Since both modules were doing this the same way, this patch
introduces a test util module to do basic versioning with only one
version, aka unversioned.

The socket test also had some chain sync tests. These are kept in the
ouroboros-network tests, and only the tests using request/response are
moved into ouroboros-network-framework.

The ouroboros-network package now no longer depends on
typed-protocol-examples at all.
