+++ ouroboros-network-framework/test/Main.hs
@@ -5,6 +5,8 @@ import           Test.Tasty
 import qualified Test.Network.TypedProtocol.PingPong.Codec as PingPong
 import qualified Test.Network.TypedProtocol.ReqResp.Codec as ReqResp
 import qualified Test.Ouroboros.Network.Driver as Driver
+import qualified Test.Ouroboros.Network.Socket as Socket
+import qualified Test.Ouroboros.Network.Subscription as Subscription
 
 main :: IO ()
 main = defaultMain tests
@@ -15,6 +17,8 @@ tests =
   [ PingPong.tests
   , ReqResp.tests
   , Driver.tests
+  , Socket.tests
+  , Subscription.tests
   ]
