+++ ouroboros-network-framework/test/Test/Ouroboros/Network/Subscription.hs
@@ -11,7 +11,7 @@
 
 {-# OPTIONS_GHC -Wno-orphans     #-}
 
-module Test.Subscription (tests) where
+module Test.Ouroboros.Network.Subscription (tests) where
 
 import           Control.Concurrent hiding (threadDelay)
 import           Control.Monad (replicateM, unless, when)
@@ -33,12 +33,6 @@ import           Data.Void (Void)
 import           Data.Word
 import qualified Network.DNS as DNS
 import qualified Network.Socket as Socket
-import           Test.QuickCheck
-import           Test.Tasty (TestTree, testGroup)
-import           Text.Printf
-import           Text.Show.Functions ()
-
-import           Test.Tasty.QuickCheck (shuffle, testProperty)
 
 --TODO: time utils should come from elsewhere
 import           Network.Mux.Time (microsecondsToDiffTime)
@@ -48,24 +42,30 @@ import qualified Network.TypedProtocol.ReqResp.Server as ReqResp
 import qualified Network.TypedProtocol.ReqResp.Codec.CBOR as ReqResp
 import qualified Network.TypedProtocol.ReqResp.Examples   as ReqResp
 
-import           Ouroboros.Network.Protocol.Handshake.Type (cborTermVersionDataCodec)
-import           Ouroboros.Network.Protocol.Handshake.Version (acceptableVersion, simpleSingletonVersions)
+import           Ouroboros.Network.Protocol.Handshake.Type
+import           Ouroboros.Network.Protocol.Handshake.Version
 
 import           Ouroboros.Network.Driver
+import           Ouroboros.Network.ErrorPolicy
 import           Ouroboros.Network.IOManager
-import           Ouroboros.Network.Magic
 import           Ouroboros.Network.Mux
-import           Ouroboros.Network.NodeToNode hiding ( ipSubscriptionWorker
-                                                     , dnsSubscriptionWorker
-                                                     )
-import           Ouroboros.Network.Snocket
 import           Ouroboros.Network.Socket
+import           Ouroboros.Network.Snocket
 import           Ouroboros.Network.Subscription
 import           Ouroboros.Network.Subscription.Dns
 import           Ouroboros.Network.Subscription.Ip
 import           Ouroboros.Network.Subscription.PeerState
 import           Ouroboros.Network.Subscription.Subscriber
-import           Ouroboros.Network.Subscription.Worker (WorkerParams (..))
+import           Ouroboros.Network.Subscription.Worker (LocalAddresses(..), WorkerParams(..))
+
+import           Test.Ouroboros.Network.Handshake
+
+import           Test.QuickCheck
+import           Test.Tasty (TestTree, testGroup)
+import           Test.Tasty.QuickCheck (testProperty, shuffle)
+import           Text.Printf
+import           Text.Show.Functions ()
+
 
 defaultMiniProtocolLimit :: Int64
 defaultMiniProtocolLimit = 3000000
@@ -558,11 +558,7 @@ prop_send_recv f xs _first = ioProperty $ withIOManager $ \iocp -> do
         (Socket.addrAddress responderAddr)
         cborTermVersionDataCodec
         (\(DictVersion _) -> acceptableVersion)
-        (simpleSingletonVersions
-          NodeToNodeV_1
-          (NodeToNodeVersionData $ NetworkMagic 0)
-          (DictVersion nodeToNodeCodecCBORTerm)
-          (SomeResponderApplication responderApp))
+        (unversionedProtocol (SomeResponderApplication responderApp))
         nullErrorPolicies
         $ \_ _ -> do
           dnsSubscriptionWorker'
@@ -584,8 +580,7 @@ prop_send_recv f xs _first = ioProperty $ withIOManager $ \iocp -> do
                 iocp
                 cborTermVersionDataCodec
                 nullNetworkConnectTracers
-                (simpleSingletonVersions NodeToNodeV_1 (NodeToNodeVersionData $ NetworkMagic 0)
-                (DictVersion nodeToNodeCodecCBORTerm) initiatorApp))
+                (unversionedProtocol initiatorApp))
 
     res <- atomically $ (,) <$> takeTMVar sv <*> takeTMVar cv
     return (res == mapAccumL f 0 xs)
@@ -697,11 +692,7 @@ prop_send_recv_init_and_rsp f xs = ioProperty $ withIOManager $ \iocp -> do
         responderAddr
         cborTermVersionDataCodec
         (\(DictVersion _) -> acceptableVersion)
-        (simpleSingletonVersions
-          NodeToNodeV_1
-          (NodeToNodeVersionData $ NetworkMagic 0)
-          (DictVersion nodeToNodeCodecCBORTerm)
-          (SomeResponderApplication (appX rrcfg)))
+        (unversionedProtocol (SomeResponderApplication (appX rrcfg)))
         nullErrorPolicies
         $ \localAddr _ -> do
           atomically $ putTMVar localAddrVar localAddr
@@ -719,11 +710,7 @@ prop_send_recv_init_and_rsp f xs = ioProperty $ withIOManager $ \iocp -> do
           responderAddr
           cborTermVersionDataCodec
           (\(DictVersion _) -> acceptableVersion)
-          (simpleSingletonVersions
-            NodeToNodeV_1
-            (NodeToNodeVersionData $ NetworkMagic 0)
-            (DictVersion nodeToNodeCodecCBORTerm)
-            (SomeResponderApplication (appX rrcfg)))
+          (unversionedProtocol (SomeResponderApplication (appX rrcfg)))
           nullErrorPolicies
           $ \localAddr _ -> do
             peerStatesVar <- newPeerStatesVar
@@ -747,8 +734,7 @@ prop_send_recv_init_and_rsp f xs = ioProperty $ withIOManager $ \iocp -> do
                   iocp
                   cborTermVersionDataCodec
                   nullNetworkConnectTracers
-                  (simpleSingletonVersions NodeToNodeV_1 (NodeToNodeVersionData $ NetworkMagic 0)
-                  (DictVersion nodeToNodeCodecCBORTerm) $ appX rrcfg))
+                  (unversionedProtocol (appX rrcfg)))
 
             atomically $ (,) <$> takeTMVar (rrcServerVar rrcfg)
                              <*> takeTMVar (rrcClientVar rrcfg)
@@ -818,8 +804,7 @@ _demo = ioProperty $ withIOManager $ \iocp -> do
                 iocp
                 cborTermVersionDataCodec
                 nullNetworkConnectTracers
-                (simpleSingletonVersions NodeToNodeV_1 (NodeToNodeVersionData $ NetworkMagic 0)
-                (DictVersion nodeToNodeCodecCBORTerm) appReq))
+                (unversionedProtocol appReq))
 
     threadDelay 130
     -- bring the servers back again
@@ -838,11 +823,7 @@ _demo = ioProperty $ withIOManager $ \iocp -> do
             (Socket.addrAddress addr)
             cborTermVersionDataCodec
             (\(DictVersion _) -> acceptableVersion)
-            (simpleSingletonVersions
-                NodeToNodeV_1
-                (NodeToNodeVersionData $ NetworkMagic 0)
-                (DictVersion nodeToNodeCodecCBORTerm)
-                (SomeResponderApplication appRsp))
+            (unversionedProtocol (SomeResponderApplication appRsp))
             nullErrorPolicies
             (\_ _ -> threadDelay delay)
