+++ ouroboros-network/test/Main.hs
@@ -8,7 +8,6 @@ import qualified Test.Chain (tests)
 import qualified Test.ChainFragment (tests)
 import qualified Test.ChainProducerState (tests)
 import qualified Test.Pipe (tests)
-import qualified Test.Subscription (tests)
 import qualified Test.PeerState (tests)
 import qualified Test.Ouroboros.Network.MockNode (tests)
 import qualified Test.Ouroboros.Network.BlockFetch (tests)
@@ -44,7 +43,6 @@ tests =
     -- network logic
   , Test.Pipe.tests
   , Test.Socket.tests
-  , Test.Subscription.tests
   , Test.PeerState.tests
   , Test.Ouroboros.Network.BlockFetch.tests
   , Ouroboros.Network.PeerSelection.Test.tests
